feat: migrates AccessDisplay to its own component
